Template generator, query string, more cards
- Finished the template generator for sending out emails. It will
create the URL with the users card code as a query string so they don't
have to enter it.
- Added the functionality to read a query string that contains the
users code to retrieve their card. If they use a URL containing a code
they will just be able to click to get the card and not have to enter
it. If an error occurs or no query string is present the text box to
enter a code will display instead.
- Added some more Christmas cards
- Edited some of the CSS for polish reasons